mkosi: add helper to reduce duplicated code
The same logic was used for dnf/tdnf/rpm/dpkg/apt, so let's split
it out to a helper.

I opted to use full paths with "/". Such paths are easier to read
and we avoid the risk of removing something from the host if somebody
uses an absolute path by mistake.