Fixed regression when using UsrOnly during initial install
Several distributions call `run_workspace_command()` during installation and
this now requires the root directory that gets bind mounted in to exist. The
created directory will be removed anyway since it exists under a temporary
workspace.

As an example, the documentation removal when installing a Debian derivative
will fail without this change.