Skip setting kernel root cmdline parameter for UsrOnly with ImageID and without verity
When an image is built as a /usr only image and does not use verity, the kernel
command line should (and already does) contain the parameter `mount.usr`. It
should not contain a root parameter since `systemd-repart` and
`systemd-gpt-auto-generator` will take care of generating and mounting the root
partition.

Without this change the root partition label used on the kernel command line
will be invalid.