Always use a pseudo-tty when boot verb is used
A container where we boot systemd is not useful without /dev/console
in it. If --pipe is used, /dev/console is not available. To make sure
/dev/console is always available when we're booting the container,
only apply the --pipe logic when booting in shell mode.

Fixes missing console output when booting a container image in a
non-interactive shell script.