Rename use_mirror_verbatim to local_mirror, support Debian/Arch/SUSE
With some distros the mirror has to be configured inside the image
(apt, zypper), so having a boolean flag is not enough.
We need to use the local mirror only for the local image build, but
configure the actual remote repositories in the image configuration,
so that it can be used once installed.