Add final cache path to MkosiState
Instead of changing the config cache path to the new temporary
cache path if no cache path is configured, let's not touch the
config and store the final cache path in the state instead.

This also changes the temporary cache path to be in the workspace
instead of in the output directory.