Use mount's new X-mount.idmap option for bind mount
Use X-mount.idmap option of mount to map current user in the new
namespace to the owner of the mounted directory on the backing
filesystem.

This prevents files and folders created in the mounted directory to be
owned by the current user once the directory is unmounted.

If mount doesn't support X-mount.idmap option, it is silently ignored
and the directory is mounted without the option.